Fix `encryptMessage` to use the specified compression algorithm (#125)
This commit changes `encryptMessage` to treat `options.compression` as an algorithm
specifier.
According to the type declarations, `options.compression` should be an
algorithm, but the library treated it as a flag. Thus, if a compression option
was specified, `zlib` would always be used for compression.

It is potentially a breaking change if the function was used as implemented,
i.e. by passing a boolean to `options.compression`.