Take `text/binaryData` as input of `encrypt/sign/verifyMessage`, remove `createMessage/CleartextMessage` (#133)
Changes:
- Remove `createMessage` and `createCleartextMessage`
- Change `encryptMessage`, `signMessage`, `verifyMessage`:
  * accept input data as either `textData` or `binaryData` (instead of `data`).
  * a `message` input can no longer be passed.
- in `signMessage` and `encryptMessage`, `textData` is used as-is. Previously,
if some string `data` was passed, `signMessage` would create a CleartextMessage
automatically, whereas in `encryptMessage` trailing spaces were automatically
removed (namely, the data was treated again as a CleartextMessage payload).

In `encryptMessage`, taking `text/binaryData` instead of a `message` is needed
to support encrypt + detached sign with streamed input data.
In `signMessage` and `verifyMessage`, the change is motivated by how the
functions are currently used in the web apps (namely, only to sign/verify detached data).
This also makes it easier to integrate the functions with a web worker, since
the input options are all of transferable type.