Add build step before publishing the package (#174)
Add `prepack` npm script that automatically runs before `npm publish`.
The script runs basic compilation from TypeScript to CommonJS and
outputs the result to the `dist` folder. This way, it will be possible
to import the pmcrypto library as a CommonJS dependency, which is
required in some contexts, like Playwright.