`argon2`: reload wasm module after timeout when using memory heavy params
We manually reload the module if no memory-heavy (128MB+) argon2 computation has been requested
in a while, to deallocate the memory.
This is better than reloading the module every time (automatically done if the memory exceeds
`Argon2S2K.ARGON2_WASM_MEMORY_THRESHOLD_RELOAD`),
as doing so comes with a performance hit.